robot_ppic/requirements.txt
supabase
requests
pytz
python-telegram-bot